+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for MiTEE-1
 - Preliminary support for CAPE-3 (only ASFK AX.25)
 - Support for EXOCUBE-2
+- PDU Length Filter block
 
 ### Changed
 - Add filter before quadrature demod for IQ input FSK demodulator